v5.0.1 patch
Fixed a bug where if the user inverted the page and then triggered a popup menu, the page would become cropped. The only way to escape this would be to reload the page, as saving work would become impossible. This bug was only apparent in the web-based editor, so users with the offline app are not affected.